refactor(modularization): Phase 4 Step 4-S6 — model_platform services + views move (#25)
- 3 services moved: llm_provider + llm_available_model + default_model → aperag/domains/model_platform/service/
- 2 views moved & renamed: views/llm.py → llm_routes.py, views/providers_v2.py → providers_v2_routes.py (under aperag/domains/model_platform/api/)
- Rewrote all view_models → domain schemas, User ORM → AuthenticatedUser(Protocol), service imports → domain paths
- G15 literal compare落地: 9 处 user.role == 'admin' 字面量（providers_v2 admin-only + publish flow）
- aperag.llm.* untouched (shared infra per Q6 / msg=d47fa490 Section 7)
- aperag.query / aperag.utils imports preserved (infrastructure, not banned)
- Legacy 5 shims (3 service + 2 view) forwarding to domain paths; views/llm.py explicit re-export for private _build_rerank_response_items used by test
- app.py include_router updated: llm_router + providers_v2_router direct from domain, same /api/v1 + /api/v2 prefixes
- OpenAPI byte-stable ✓
- 671 unit tests pass
- 15th commit on bryce/phase4-impl
